requirements/common.txt
Django==1.8.2
MySQL-python==1.2.5
argparse==1.2.1
django-allauth==0.20.0
django-bootstrap3==6.1.0
django-braces==1.8.0
djrill==1.4.0
gunicorn==19.3.0
oauthlib==0.7.2
python-openid==2.2.5
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wsgiref==0.1.2
